Functions/Cloud Functions/Restaurant_polarity/requirements.txt
google-cloud-firestore>=2.1.3
google-cloud-storage>=2.6.0
pandas>=1.5.2